% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_dat_input}
\alias{check_dat_input}
\title{Check dat input parameter}
\usage{
check_dat_input(dat, geo_var, values_var)
}
\arguments{
\item{dat}{Eurostat data frame or a data frame derived from
\code{create_indicator} or \code{create_tables}.}

\item{geo_var}{The name of the variable that contains the standard
\code{geo} codes. It defaults to \code{"geo"}.}

\item{values_var}{The name of the variable that contains the
\code{values} To be mapped. It defaults to \code{"values"}. If the data
you want to see on map is in the column \code{average_values}, than
use  \code{"average_values"}. The values can be numeric or categorical
variables.}
}
\value{
The function does not return a value, but creates a meaningful
error message if the dat parameter is non-conforming.
}
\description{
Check dat input parameter
}
\keyword{internal}
