% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_base_plot_cat}
\alias{create_base_plot_cat}
\title{Create a categorical choropleth}
\usage{
create_base_plot_cat(
  choropleth_data,
  color_palette,
  iceland,
  are_there_missings,
  unit_text
)
}
\arguments{
\item{choropleth_data}{An sf object with the categorical data in column
\code{cat}.}

\item{color_palette}{A named character vector with colors. If you use this
with categorical variables, make sure that the color_palette has a value
for each categories except for missing values, and it is named to the
category name (factor level) as it is found in you \code{values_var}.}

\item{iceland}{A logical variable to show Iceland on the map.}

\item{are_there_missings}{A logical variable if missing data is present.}

\item{unit_text}{Defaults to \code{NULL}. The title caption of the
color (fill) legend.}
}
\value{
A ggplot object containing the choropleth map.
}
\description{
Create a categorical choropleth
}
\keyword{internal}
