% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_choropleth.R
\name{create_choropleth}
\alias{create_choropleth}
\title{Create A Choropleth Map}
\usage{
create_choropleth(
  dat,
  n = 5,
  geo_var = "geo",
  values_var = "values",
  type = "discrete",
  show_all_missing = TRUE,
  unit_text = NULL,
  color_palette = NULL,
  na_color = "grey93",
  drop_levels = FALSE,
  reverse_scale = FALSE,
  style = "pretty",
  print_style = "min-max",
  iceland = "if_present"
)
}
\arguments{
\item{dat}{Eurostat data frame or a data frame derived from
\code{create_indicator} or \code{create_tables}.}

\item{n}{Number of colour categories if discrete (categorical) colouring
is chosen, defaults to \code{5}.}

\item{geo_var}{The name of the variable that contains the standard
\code{geo} codes. It defaults to \code{"geo"}.}

\item{values_var}{The name of the variable that contains the
\code{values} To be mapped. It defaults to \code{"values"}. If the data
you want to see on map is in the column \code{average_values}, than
use  \code{"average_values"}. The values can be numeric or categorical
variables.}

\item{type}{Defaults to \code{'discrete'} for coloring discrete
(categorical) values, can be set to \code{'numeric'} for continous
numeric values.}

\item{show_all_missing}{Defaults to \code{TRUE} when all regions with
missing values are explicitly shown in the map.  \code{FALSE} shows only
missing observations in \code{dat}. For example, if your dataset contains
observations on European Union regions, than it will not contain an
\code{NA} value or Norwegian regions in the case of \code{FALSE}.}

\item{unit_text}{Defaults to \code{NULL}. The title caption of the
color (fill) legend.}

\item{color_palette}{A named character vector with colors. If you use this
with categorical variables, make sure that the color_palette has a value
for each categories except for missing values, and it is named to the
category name (factor level) as it is found in you \code{values_var}.}

\item{na_color}{Defaults to \code{"grey93"}. This color will be used
in the color palette for missing values, unless you have explicitly
set one of the names of the palette to \code{"missing"}.}

\item{drop_levels}{Weather to drop categorical levels on the choropleth
if they are not present in the data.  Defaults to \code{FALSE}.}

\item{reverse_scale}{Defaults to \code{FALSE}.}

\item{style}{Style of interval reporting, defaults to \code{"pretty"}.
Style can be one of \code{"fixed"}, \code{"sd"}, \code{"equal"},
\code{"pretty"}, \code{"quantile"}, \code{"kmeans"}, \code{"hclust"},
\code{"bclust"}, \code{"fisher"}, \code{"jenks"} or \code{"dpih"}.}

\item{print_style}{Style of printing the labels, defaults to
\code{"min-max"}. Alternative is \code{"interval"}}

\item{iceland}{Show Iceland on the choropleth?  Defaults to
\code{"if_present"}, which shows Iceland if Icelandic data
is present.
See: \code{\link[classInt]{classIntervals}}.}
}
\value{
The function returns a \code{ggplot2} object. You can modify the
ggplot object, for example, with adding {labs}.
}
\description{
This function will create a choropleth map of the European Union and some
of its (perspective) member candidates, the EEA countries, and the United
Kingdom. The geographical information must conform the official \code{geo}
codes of Eurostat.
}
\examples{
\dontrun{
chreate_choropleth ( your_nuts2_data, level=2, n=5, style="kmeans") +
 labs (title = "Your Title", fill = "fill legend name",
       subtitle = "Your Subtitle", caption ="Your caption or footnote.")
}
}
\seealso{
Other visualisation functions: 
\code{\link{create_color_palette}()},
\code{\link{sr_palette}()}
}
\concept{visualisation functions}
