% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_color_palette.R
\name{create_color_palette}
\alias{create_color_palette}
\title{Create A Discrete Color Palette}
\usage{
create_color_palette(
  base_color = "#00348A",
  min_color = "grey90",
  n = n,
  reverse_scale = FALSE
)
}
\arguments{
\item{base_color}{Defaults to \code{"#00348A"}.}

\item{min_color}{Defaults to \code{"grey90"}.}

\item{n}{Number of colors, defaults to \code{5}.}

\item{reverse_scale}{Defaults to \code{FALSE}.}
}
\description{
This is a wrapper around \code{\link[grDevices]{colorRampPalette}}.
The default colors are taken from \code{\link{sr_palette}}.
}
\examples{
{
create_color_palette(base_color = 'red', min_color = 'white', n=3)
create_color_palette('red', 'white', n=4, reverse_scale =TRUE)
}
}
\seealso{
sr_palette
}
