% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_information.R
\docType{data}
\name{geodata_nuts1}
\alias{geodata_nuts1}
\title{NUTS1 level boundary information}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 125 rows and 8 columns.
}
\usage{
geodata_nuts1
}
\description{
NUTS1 level boundary information
}
\keyword{datasets}
