% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_information.R
\docType{data}
\name{geodata_nuts2}
\alias{geodata_nuts2}
\title{NUTS2 level boundary information}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 332 rows and 8 columns.
}
\usage{
geodata_nuts2
}
\description{
NUTS2 level boundary information
}
\keyword{datasets}
