% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_categories.R
\name{indicator_categories}
\alias{indicator_categories}
\title{Create Categories From Numeric Vector.}
\usage{
indicator_categories(
  values,
  n = 5,
  dataPrecision = 0,
  style = "pretty",
  print_style = "min-max"
)
}
\arguments{
\item{values}{A numeric vector that may contain missing values.}

\item{n}{Number o intervals, defaults to \code{5}.}

\item{dataPrecision}{= Number of digits on labels, defaults to
\code{0}.}

\item{style}{Style of interval reporting, defaults to \code{"pretty"}.
Style can be one of \code{"fixed"}, \code{"sd"}, \code{"equal"},
\code{"pretty"}, \code{"quantile"}, \code{"kmeans"}, \code{"hclust"},
\code{"bclust"}, \code{"fisher"}, \code{"jenks"} or \code{"dpih"}.
See: \code{\link[classInt]{classIntervals}}.}

\item{print_style}{Style of printing the labels, defaults to
\code{"min-max"}. Alternative is \code{"interval"}}
}
\value{
A character vector containing the category labels.
}
\description{
A wrapper around \code{\link[classInt]{classIntervals}} that
creates printable univariate class intervals from numeric values.
}
\examples{
{
indicator_categories(1:7, n=2)
indicator_categories(c(1:6, NA), n=2, print_style = "interval")
}
}
\concept{utility functions}
