% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nuts_example_missings.R
\docType{data}
\name{nuts_with_missings}
\alias{nuts_with_missings}
\title{Random Data With NUTS2016 Coding}
\format{
A data frame with 900 rows and 2 variables:
\describe{
  \item{geo}{NUTS geo identifier}
  \item{values}{A random number.}
}
}
\usage{
nuts_with_missings
}
\description{
Random Data With NUTS2016 Coding
}
\keyword{datasets}
