% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_europe_countries.R
\name{palette_europe_countries}
\alias{palette_europe_countries}
\title{Return a color for each European country from the SR palette.}
\usage{
palette_europe_countries(europe_countries_4 = FALSE)
}
\arguments{
\item{eu_countries_4}{If \code{TRUE}, returns values for \code{GB-G},
\code{GB-N}, \code{DE-W}, \code{DE-E}.}
}
\value{
A named vector of color hex strings for each country code.
}
\description{
Return a color for each European country from the SR palette.
}
\examples{
palette_europe_countries( europe_countries_4 = TRUE)
}
\seealso{
Other palette constructor functions: 
\code{\link{create_color_palette}()},
\code{\link{palette_eu_countries}()},
\code{\link{sr_palette}()}
}
\concept{palette constructor functions}
