% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satellitereport.R
\docType{package}
\name{satellitereport}
\alias{satellitereport}
\title{Satellite Report Utilities}
\description{
This package helps cooperation with Satellite Report. \url{http://satellitereport.com}
}
\section{utility functions}{

\code{\link{indicator_categories}} is a wrapper around
\code{\link[classInt]{classIntervals}} and
creates printable univariate class intervals from numeric values.
}

\section{visualisation functions}{


\code{\link{create_color_palette}} creates a gradual discrete color
palette.

\code{\link{sr_palette}} returns the satellitereport.com
color palette.

\code{\link{create_choropleth}} will create a ggplot2 object with a map
of European regions and continous or categorical data visualised on it. It
can convert continous (numeric) data into categories.
}

