% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sr_palette.R
\name{sr_palette}
\alias{sr_palette}
\title{SR Color Palette}
\usage{
sr_palette(ext = FALSE)
}
\arguments{
\item{ext}{Should the original 12 color palette be extended? Defaults
to \code{FALSE}.}
}
\value{
The Satellit Report color palette designed by Endre Koronczi.
}
\description{
SR Color Palette
}
\examples{
{
sr_palette()
sr_palette(ext=TRUE)
}
}
\seealso{
Other palette constructor functions: 
\code{\link{create_color_palette}()},
\code{\link{palette_eu_countries}()},
\code{\link{palette_europe_countries}()}
}
\concept{palette constructor functions}
