% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_wrap.R
\name{text_wrap}
\alias{text_wrap}
\title{Wrap Text To Adequate Width}
\usage{
text_wrap(text, width = 50)
}
\arguments{
\item{text}{A character vector that is too long to fit in a caption.}

\item{width}{Number of characters to allow in text,
defaults to \code{50}.}
}
\value{
A character vector containing the category labels.
}
\description{
A wrapper around \code{\link[base]{strwrap}} that
creates wrapped text for captions.
}
\examples{
{
my_text <- "Regional gross domestic product by NUTS 2 regions - million EUR"
nchar(my_text)
text_wrap (my_text, 40)
}
}
\seealso{
Other utility functions: 
\code{\link{indicator_categories}()}
}
\concept{utility functions}
