% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-europe_shapefile.R
\docType{data}
\name{geodata_europe_2016}
\alias{geodata_europe_2016}
\title{The sf File Of NUTS Boundaries in Europe (NUTS2016)}
\format{
A data frame with 1951 observations of 3 variables:
\describe{
  \item{NUTS_ID}{The NUTS geocodes}
  \item{STAT_LEVL_}{The NUTS level code, 0,1,2 or 3.}
  \item{geometry}{A sfc_MULTIPOLYGON of the European country boundaries}
  }
}
\source{
\url{https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data/administrative-units-statistical-units/nuts}
}
\usage{
geodata_europe_2016
}
\description{
The sf File Of NUTS Boundaries in Europe (NUTS2016)
}
\keyword{datasets}
