% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundary_information.R
\docType{data}
\name{geodata_nuts3}
\alias{geodata_nuts3}
\title{NUTS0 level boundary information}
\format{
An object of class \code{sf} (inherits from \code{data.frame}) with 1522 rows and 8 columns.
}
\usage{
geodata_nuts3
}
\description{
NUTS0 level boundary information
}
\keyword{datasets}
